% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ped2num.R
\name{ped2num}
\alias{ped2num}
\title{Function for transforming genomic data (GeneSeek 50K (Illumina) SNP chip) from nucleobases (A,C,G,T) to 0-1-2 coding}
\usage{
ped2num(ped, na_val = NA)
}
\arguments{
\item{ped:}{A data frame of  size n x 2m, where n is the number of rows (individuals)
in the ped file, and m is the number of SNP's in the map file (23 070 in the PigBreedPrediction_map)}

\item{na_val:}{Value to put in SNP's with missing data. Default is NA, might be set to 0.}
}
\value{
matrix size n x m, where n is the number of rows (individuals)
in the ped file, and m is the number of SNP's in the map file
(23 070 in the PigBreedPrediction_map), use
head(PigBreedPrediction_map) to view first entries of map file. Column names
according to SNP-names and Rownames according to ID
}
\description{
Function for transforming genomic data (GeneSeek 50K (Illumina) SNP chip) from nucleobases (A,C,G,T) to 0-1-2 coding
}
\examples{
ped <- read_ped_data('data/Ped_example.txt',ID_column = 'ID',nrows = -1)
ped012 <- ped2num(ped,na_val = 0)
}
\references{
Vinje H, Brustad HK, Heggli A, Sevillano CA, Van Son M,
Gangsei LE. Classification of breed combinations for slaughter pigs based
on genotypes-modeling DNA samples of crossbreeds as fuzzy sets from purebred
founders. Front Genet. 2023 Dec 4;14:1289130. doi: 10.3389/fgene.2023.1289130.
PMID: 38116292; PMCID: PMC10729766.
}
\author{
Hilde Vinje & Lars Erik Gangsei
}
